fix: wrap `emitSync` calls in `setImmediate`
This prevents reentrancy in the following scenario:
If a listener changes a file and then blocks the event loop (or the next listener blocks it), and multiple file events were batched, those file events may be processed *after* the event from the previously mentioned listener.

Now, an event batch is processed entirely before the emit queue is processed.